Task 12
Completed changes to PropertyCell.java methods for method refactoring. I
chose not to make the String array a parameter- I did not think it was
necessary to send the array when calling the new method; moreover, the
array is now in a private method, and thus the data is better protected
in this class.